utxo: Move UTXO code to it's own file.
